% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\docType{data}
\name{ANPC_EXCLUDE_PATTERN}
\alias{ANPC_EXCLUDE_PATTERN}
\title{PeakForgeR_mzml.R}
\format{
An object of class \code{character} of length 1.
}
\usage{
ANPC_EXCLUDE_PATTERN
}
\description{
mzML file import, parsing, and chromatogram handling functions.
Includes mzR MRM finding, peak boundary detection, and lipid matching.
}
\keyword{datasets}
