% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_archive.R
\name{archive_files}
\alias{archive_files}
\title{Archive Files}
\usage{
archive_files(project_directory, folder_name)
}
\arguments{
\item{project_directory}{The directory of the project.}

\item{folder_name}{The name of the folder to be archived.}
}
\value{
invisible(NULL)
}
\description{
This function moves a specified folder to the archive directory within the project directory.
}
\examples{
\dontrun{
archive_files(project_directory, folder_name)
}
}
\keyword{internal}
