% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_dataprep.R
\name{assess_qc_coverage}
\alias{assess_qc_coverage}
\title{Assess QC Coverage}
\usage{
assess_qc_coverage(master_list)
}
\arguments{
\item{master_list}{A list containing project details and sorted peak area data.}
}
\value{
The updated master list with QC coverage results.
}
\description{
This function assesses the QC coverage for each plate in the master list.
It checks the ratio of QC samples to total samples and determines if the QC passed or failed.
}
\keyword{internal}
