% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_adjust_corrected_means}
\alias{bc_adjust_corrected_means}
\title{Adjust Corrected QC Means to Original Scale}
\usage{
bc_adjust_corrected_means(
  corrected_clean,
  data_flagged,
  qc_label,
  metabolite_cols
)
}
\arguments{
\item{corrected_clean}{Tibble of corrected data.}

\item{data_flagged}{Original (flagged) data.frame.}

\item{qc_label}{Character identifying QC samples.}

\item{metabolite_cols}{Character vector of metabolite column names.}
}
\value{
Tibble with mean-adjusted corrected values.
}
\description{
Adjust Corrected QC Means to Original Scale
}
\keyword{internal}
