% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_calculate_rsd}
\alias{bc_calculate_rsd}
\title{Calculate Percent RSD for QC Samples}
\usage{
bc_calculate_rsd(data, qc_label, metabolite_cols)
}
\arguments{
\item{data}{A data.frame of sample data.}

\item{qc_label}{Character identifying QC samples.}

\item{metabolite_cols}{Character vector of metabolite column names.}
}
\value{
Named numeric vector of percent RSD values.
}
\description{
Calculate Percent RSD for QC Samples
}
\keyword{internal}
