% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_clean_correction_output}
\alias{bc_clean_correction_output}
\title{Clean statTarget Correction Output}
\usage{
bc_clean_correction_output(corrected_raw, metabolite_map, pheno)
}
\arguments{
\item{corrected_raw}{Tibble of raw statTarget output.}

\item{metabolite_map}{Tibble mapping metabolite codes to original names.}

\item{pheno}{Tibble with sample ID to sample_name mapping.}
}
\value{
Tibble with sample_name and corrected metabolite columns.
}
\description{
Clean statTarget Correction Output
}
\keyword{internal}
