% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_detect_metabolite_columns}
\alias{bc_detect_metabolite_columns}
\title{Detect Metabolite Columns}
\usage{
bc_detect_metabolite_columns(data)
}
\arguments{
\item{data}{A data.frame containing sample and metabolite data.}
}
\value{
Character vector of metabolite column names.
}
\description{
Detect Metabolite Columns
}
\keyword{internal}
