% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_flag_failed_qc}
\alias{bc_flag_failed_qc}
\title{Flag Failed QC Injections}
\usage{
bc_flag_failed_qc(data, qc_label, metabolite_cols)
}
\arguments{
\item{data}{A data.frame of sample data.}

\item{qc_label}{Character identifying QC samples.}

\item{metabolite_cols}{Character vector of metabolite column names.}
}
\value{
List with \code{data} (modified) and \code{failed_samples} (character).
}
\description{
Flag Failed QC Injections
}
\keyword{internal}
