% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_plot_correction_results}
\alias{bc_plot_correction_results}
\title{Generate Correction Visualisations}
\usage{
bc_plot_correction_results(
  original_data,
  corrected_data,
  qc_label,
  metabolite_cols,
  qc_rsd_before,
  qc_rsd_after
)
}
\arguments{
\item{original_data, corrected_data}{Data.frames before/after correction.}

\item{qc_label}{Character identifying QC samples.}

\item{metabolite_cols}{Character vector of metabolite column names.}

\item{qc_rsd_before, qc_rsd_after}{Named numeric vectors of QC RSD.}
}
\value{
Named list of ggplot objects.
}
\description{
Generate Correction Visualisations
}
\keyword{internal}
