% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_plot_pca}
\alias{bc_plot_pca}
\title{Generate PCA Score Plots}
\usage{
bc_plot_pca(original_data, corrected_data, metabolite_cols)
}
\arguments{
\item{original_data, corrected_data}{Data.frames before/after correction.}

\item{metabolite_cols}{Character vector of metabolite column names.}
}
\value{
A ggplot object or NULL.
}
\description{
Generate PCA Score Plots
}
\keyword{internal}
