% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_prepare_pheno_file}
\alias{bc_prepare_pheno_file}
\title{Prepare PhenoFile for statTarget}
\usage{
bc_prepare_pheno_file(data, qc_label, st_dir)
}
\arguments{
\item{data}{Data.frame with flagged sample types.}

\item{qc_label}{Character identifying QC samples.}

\item{st_dir}{Path to statTarget working directory.}
}
\value{
Tibble with pheno mapping columns: sample, sample_name, batch, class, order.
}
\description{
Prepare PhenoFile for statTarget
}
\keyword{internal}
