% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_prepare_profile_file}
\alias{bc_prepare_profile_file}
\title{Prepare ProfileFile for statTarget}
\usage{
bc_prepare_profile_file(data, metabolite_cols, pheno, st_dir)
}
\arguments{
\item{data}{Data.frame with flagged sample types.}

\item{metabolite_cols}{Character vector of metabolite column names.}

\item{pheno}{Tibble returned by \code{bc_prepare_pheno_file}.}

\item{st_dir}{Path to statTarget working directory.}
}
\value{
List with \code{metabolite_map} tibble.
}
\description{
Prepare ProfileFile for statTarget
}
\keyword{internal}
