% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_reconstruct_output}
\alias{bc_reconstruct_output}
\title{Reconstruct Output in Original Data Format}
\usage{
bc_reconstruct_output(corrected_adjusted, original_data, metabolite_cols)
}
\arguments{
\item{corrected_adjusted}{Tibble with corrected metabolite values.}

\item{original_data}{The original input data.frame.}

\item{metabolite_cols}{Character vector of metabolite column names.}
}
\value{
Tibble matching the structure of the original data.
}
\description{
Reconstruct Output in Original Data Format
}
\keyword{internal}
