% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_run_batch_correction}
\alias{bc_run_batch_correction}
\title{Run statTarget Batch Correction}
\usage{
bc_run_batch_correction(st_dir, method, ntree, coCV, Frule, imputeM)
}
\arguments{
\item{st_dir}{Path to statTarget working directory.}

\item{method, ntree, coCV, Frule, imputeM}{See \code{batchCorrectR}.}
}
\value{
Tibble of raw corrected data from statTarget.
}
\description{
Run statTarget Batch Correction
}
\keyword{internal}
