% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrectR_Utils.R
\name{bc_validate_input}
\alias{bc_validate_input}
\title{Validate batchCorrectR Input}
\usage{
bc_validate_input(data, qc_label, method, ntree, coCV, Frule, imputeM)
}
\arguments{
\item{data, qc_label, method, ntree, coCV, Frule, imputeM}{See \code{batchCorrectR}.}
}
\value{
Invisible TRUE if all checks pass.
}
\description{
Validate batchCorrectR Input
}
\keyword{internal}
