% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_filter.R
\name{calculate_lipid_flags}
\alias{calculate_lipid_flags}
\title{Calculate lipid flags for a batch}
\usage{
calculate_lipid_flags(master_list, idx_batch, lipid_matrix)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{idx_batch}{The index of the batch (plate) to process.}

\item{lipid_matrix}{A matrix containing lipid data for the batch.}
}
\value{
A tibble containing lipid flags, including counts of peak areas below a threshold, missing values, and flags for excessive missing values.
}
\description{
This function calculates flags for lipids based on their signal intensity and missing values in a specific batch.
It checks for SIL internal standards, counts peak areas below a threshold, and flags plates with excessive missing values.
}
\keyword{internal}
