% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{calculate_runtime}
\alias{calculate_runtime}
\title{Calculate Runtime}
\usage{
calculate_runtime(master_list, section_name, previous_section_name)
}
\arguments{
\item{master_list}{A list containing project details and script log information.}

\item{section_name}{A string representing the name of the current section.}

\item{previous_section_name}{A string representing the name of the previous section.}
}
\value{
The updated \code{master_list} with the calculated runtime for the section.
}
\description{
This function calculates the runtime for the given section.
}
\keyword{internal}
