% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_filter.R
\name{calculate_sil_flags_per_plate}
\alias{calculate_sil_flags_per_plate}
\title{Calculate SIL Flags per Plate}
\usage{
calculate_sil_flags_per_plate(master_list, idx_batch)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{idx_batch}{The index of the batch (plate) to process.}
}
\value{
A tibble containing SIL flags for each lipid, including counts of peak areas below a threshold, missing values, and flags for excessive missing values.
}
\description{
This function calculates flags for SIL internal standards on a per-plate basis.
It computes the number of peak areas below a threshold, counts missing values, and flags plates with excessive missing values.
}
\keyword{internal}
