% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{check_docker}
\alias{check_docker}
\title{Function to check Docker installation, daemon, and containers}
\usage{
check_docker(auto_pull = interactive())
}
\value{
Returns current status of docker
}
\description{
Function to check Docker installation, daemon, and containers
}
\examples{
\dontrun{
check_docker()
}
}
\keyword{internal}
