% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_archive.R
\name{delete_source_directory}
\alias{delete_source_directory}
\title{delete_source_directory
This function deletes the source directory after files have been moved.}
\usage{
delete_source_directory(source_dir)
}
\arguments{
\item{source_dir}{Directory path for the folder to delete.}
}
\value{
None. The function deletes the source directory.
}
\description{
delete_source_directory
This function deletes the source directory after files have been moved.
}
\examples{
\dontrun{
delete_source_directory(source_dir = "path/to/source")
}
}
\keyword{internal}
