% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_docker.R
\name{execute_PeakForgeR_command}
\alias{execute_PeakForgeR_command}
\title{execute_PeakForgeR_command}
\usage{
execute_PeakForgeR_command(master_list, plate_idx)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{plate_idx}{The index of the plate to execute the Skyline command for.}
}
\value{
Executes the Skyline command and generates reports and chromatogram files saving to project directory.
}
\description{
This function executes a Skyline system command to process mzML files and generate various reports for a given plate.
}
\examples{
\dontrun{
execute_PeakForgeR_command(master_list, plate_idx)
}
}
\keyword{internal}
