% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_export.R
\name{export_html_report}
\alias{export_html_report}
\title{Export HTML Report
This function exports the \code{master_list} data to an HTML report using a predefined R Markdown template.
It renders the report and opens it in the default web browser.}
\usage{
export_html_report(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with the HTML report exported.
}
\description{
Export HTML Report
This function exports the \code{master_list} data to an HTML report using a predefined R Markdown template.
It renders the report and opens it in the default web browser.
}
\keyword{internal}
