% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{extract_acquisition_year}
\alias{extract_acquisition_year}
\title{extract_acquisition_year}
\usage{
extract_acquisition_year(timestamp)
}
\arguments{
\item{timestamp}{A string representing the timestamp.}
}
\value{
A numeric value representing the year extracted from the timestamp.
}
\description{
This function extracts the year from a timestamp string and converts it to a numeric value.
}
\examples{
\dontrun{
extract_acquisition_year("2025-06-17T13:31:58Z")
}
}
\keyword{internal}
