% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_dataprep.R
\name{extract_sample_id}
\alias{extract_sample_id}
\title{extract_sample_id}
\usage{
extract_sample_id(filenames)
}
\arguments{
\item{filenames}{sample_name column of .data}
}
\value{
sample_ID column with unique sample identifiers
}
\description{
This function pulls sampleID from file name.
}
\keyword{internal}
