% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{filter_mzML_filelist_qc}
\alias{filter_mzML_filelist_qc}
\title{filter_mzML_filelist_qc}
\usage{
filter_mzML_filelist_qc(mzML_filelist, FUNC_OPTION_qc_type)
}
\arguments{
\item{mzML_filelist}{A list containing mzML file information from plate.}

\item{FUNC_OPTION_qc_type}{QC type used in the experiment parsed internally.}
}
\value{
A list containing quality control sample data.
}
\description{
This function filters the mzML_filelist for user selected quality control samples
}
\examples{
\dontrun{
filter_mzML_filelist_qc(mzML_filelist,FUNC_OPTION_qc_type)
}
}
\keyword{internal}
