% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{find_lipid_info}
\alias{find_lipid_info}
\title{find_lipid_info}
\usage{
find_lipid_info(
  FUNC_mrm_guide,
  precursor_mz,
  product_mz,
  mzml_rt_apex,
  FUNC_mzR,
  idx_plate,
  idx_mzML,
  idx_mrm,
  strict_mz_tol = 0.01,
  fallback_mz_tol = 0.05,
  rt_tol = 0.1
)
}
\arguments{
\item{FUNC_mrm_guide}{Tibble of mrm details parsed internally. See run mrm_template_guide for example.}

\item{precursor_mz}{A numeric value for precursor mass to charge ratio.}

\item{product_mz}{A numeric value for product mass to charge ratio.}

\item{mzml_rt_apex}{A numeric value for retention time apex.}

\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}

\item{idx_plate}{A string naming the current plate ID being processed. Passed from process_files function.}

\item{idx_mzML}{A string naming the current mzML file being processed. Passed from process_files function.}

\item{idx_mrm}{A string identifying the current mrm. Passed from process_files function.}
}
\value{
A list contain lipid class and lipid species information.
}
\description{
This function finds and matches lipids to mrm data.
}
\examples{
\dontrun{
find_lipid_info(FUNC_mrm_guide, precursor_mz, product_mz,
                mzml_rt_apex, FUNC_mzR, idx_plate, idx_mzML, idx_mrm)
}
}
\keyword{internal}
