% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_dataprep.R
\name{find_matching_report}
\alias{find_matching_report}
\title{Find Matching PeakForgeR Report Name}
\usage{
find_matching_report(report_list, plate_id)
}
\arguments{
\item{report_list}{A list of PeakForgeR reports.}

\item{plate_id}{The ID of the plate to match.}
}
\value{
The name of the matching report, or NULL if not found.
}
\description{
This function finds the matching PeakForgeR report name based on the plate ID.
}
\keyword{internal}
