% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{find_peak_apex_idx}
\alias{find_peak_apex_idx}
\title{find_peak_apex_idx}
\usage{
find_peak_apex_idx(FUNC_mzR, idx_plate, idx_mzML, idx_mrm)
}
\arguments{
\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}

\item{idx_plate}{A string naming the current plate ID being processed. Passed from process_files function.}

\item{idx_mzML}{A string naming the current mzML file being processed. Passed from process_files function.}

\item{idx_mrm}{A string identifying the current mrm. Passed from process_files function.}
}
\value{
A numeric value for peak apex index.
}
\description{
This function finds peak apex for a single transition.
}
\examples{
\dontrun{
find_peak_apex_idx(FUNC_mzR, idx_plate, idx_mzML, idx_mrm)
}
}
\keyword{internal}
