% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_filter.R
\name{get_lipid_data}
\alias{get_lipid_data}
\title{Extract lipid matrix for a batch}
\usage{
get_lipid_data(master_list, idx_batch)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{idx_batch}{The index of the batch (plate) to process.}
}
\description{
This function extracts the lipid data matrix for a specific batch from the \code{master_list}.
It filters out samples with missing values and selects relevant columns.
}
\keyword{internal}
