% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{get_mzML_filelist}
\alias{get_mzML_filelist}
\title{get_mzML_filelist}
\usage{
get_mzML_filelist(FUNC_mzR)
}
\arguments{
\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}
}
\value{
A list containing mzML file information from plate.
}
\description{
This function generates the mzML_filelist from FUNC_mzR
}
\examples{
\dontrun{
get_mzML_filelist(FUNC_mzR)
}
}
\keyword{internal}
