% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_dataprep.R
\name{harmonise_lipid_columns}
\alias{harmonise_lipid_columns}
\title{Harmonise Lipid Columns}
\usage{
harmonise_lipid_columns(master_list)
}
\arguments{
\item{master_list}{A list containing project details and response/concentration data.}
}
\value{
The updated master list with harmonised lipid columns.
}
\description{
This function harmonises the lipid columns across response and concentration data types in the master list.
It ensures that all lipid subtypes have the same columns by selecting only the common lipids across all plates.
}
\keyword{internal}
