% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_dataprep.R
\name{lgw_impute}
\alias{lgw_impute}
\title{LGW Impute}
\usage{
lgw_impute(x)
}
\arguments{
\item{x}{A tibble containing numeric data.}
}
\value{
A tibble with zero values replaced by half the minimum non-zero value in each column.
}
\description{
This function performs imputation on a tibble by replacing zero values with half the minimum non-zero value in each column.
It calculates the minimum non-zero value for each column, divides it by 2, and replaces zero values with this calculated value.
}
\keyword{internal}
