% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{log_error}
\alias{log_error}
\title{Log Error to File}
\usage{
log_error(error_message, plateID, project_directory = getwd())
}
\arguments{
\item{error_message}{A character string representing the error message to be logged.}

\item{plateID}{A character string identifying the plate.}

\item{project_directory}{A character string for the project directory path. Defaults to \code{getwd()}.}
}
\value{
None. The function writes the error message to the log file.
}
\description{
This function logs error messages to a file named \code{error_log.txt}.
}
\examples{
\dontrun{
log_error("An error occurred while processing the data.")
}
}
\keyword{internal}
