% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_construct_command_for_terminal}
\alias{msConvertR_construct_command_for_terminal}
\title{msConvertR_construct_commands_for_terminal}
\usage{
msConvertR_construct_command_for_terminal(
  input_directory,
  output_directory,
  plateIDs,
  sanitized_plateIDs = plateIDs
)
}
\arguments{
\item{input_directory}{path to input directory containing vendor files}

\item{output_directory}{path to output directory.}

\item{plateIDs}{The names of vendor files to convert}
}
\value{
The constructed command string.
}
\description{
This function constructs the command for terminal to convert files to mzML format.
}
\examples{
\dontrun{
command <- msConvertR_construct_command_for_terminal(path/to/input/directory,
                                                     "path/to/output_directory")
}
}
\keyword{internal}
