% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_restructure_directory}
\alias{msConvertR_restructure_directory}
\title{msConvertR_restructure_directory}
\usage{
msConvertR_restructure_directory(
  output_directory,
  plateIDs,
  vendor_extension_patterns
)
}
\arguments{
\item{output_directory}{Output directory where the mzML files will be stored.}

\item{plateIDs}{filenames for plates being converted with no extension.}

\item{vendor_extension_patterns}{vector of file extensions for vendor files}
}
\value{
\code{invisible(NULL)}. Called for its side effects.
}
\description{
This function restructures the directory by moving raw_data and mzML files to correct locations.
}
\examples{
\dontrun{
master_list <- msConvertR_restructure_directory(output_directory,
                                                plateIDs,
                                                vendor_extension_patterns)
}
}
\keyword{internal}
