% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{mzR_mrm_findR}
\alias{mzR_mrm_findR}
\title{mzR_mrm_findR}
\usage{
mzR_mrm_findR(FUNC_mzR, FUNC_mrm_guide, FUNC_OPTION_qc_type)
}
\arguments{
\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}

\item{FUNC_mrm_guide}{Tibble of mrm details parsed internally. See run mrm_template_guide for example.}

\item{FUNC_OPTION_qc_type}{QC type used in the experiment passed internally.}
}
\value{
A list containing updated mrm guide and peak boundary information.
}
\description{
This function processes mzML files to find peak apex and boundaries using QC mzR data, updates the mrm guide, and provides peak boundary information.
}
\examples{
\dontrun{
mzR_mrm_findR(FUNC_mzR, FUNC_mrm_guide, FUNC_OPTION_qc_type)
}
}
\keyword{internal}
