% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_filter.R
\name{notify_qc_type}
\alias{notify_qc_type}
\title{Notify QC Type}
\usage{
notify_qc_type(qc_type)
}
\arguments{
\item{qc_type}{A string indicating the QC type (e.g. "pqc" or "ltr").}
}
\description{
This function prints a message indicating the QC type that has been set for filtering.
}
\keyword{internal}
