% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR.R, R/PeakForgeR_Utils.R
\name{PeakForgeR_import_external_functions}
\alias{PeakForgeR_import_external_functions}
\title{Import specific functions from packages}
\description{
Import specific functions from packages

Utils_Global.R
Import specific functions from packages
}
\keyword{internal}
