% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_dataprep.R
\name{prepare_imputation_matrix}
\alias{prepare_imputation_matrix}
\title{Prepare Imputation Matrix}
\usage{
prepare_imputation_matrix(data)
}
\arguments{
\item{data}{A tibble containing the sorted peak area data.}
}
\value{
A tibble ready for imputation, with problematic values replaced.
}
\description{
This function prepares the imputation matrix from the sorted peak area data.
It removes sample name columns, converts the data to a matrix, and replaces problematic values (zeros, infinite, NaN).
}
\keyword{internal}
