% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_filter.R
\name{process_lipid_versions}
\alias{process_lipid_versions}
\title{Process lipid flags across template versions}
\usage{
process_lipid_versions(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\description{
This function processes lipid flags across different template versions.
It aggregates lipid flags from all plates for each version, counts missing values, and flags versions with excessive missing values.
}
\keyword{internal}
