% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{process_mrm_transitions}
\alias{process_mrm_transitions}
\title{process_mrm_transitions}
\usage{
process_mrm_transitions(FUNC_mzR, FUNC_mrm_guide, idx_plate, idx_mzML)
}
\arguments{
\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}

\item{FUNC_mrm_guide}{Tibble of mrm details parsed internally. See run mrm_template_guide for example.}

\item{idx_plate}{A string naming the current plate ID being processed. Passed from process_files function.}

\item{idx_mzML}{A string naming the current mzML file being processed. Passed from process_files function.}
}
\value{
A tibble of mrm data for the specific idx_plate/idx_mzML.
}
\description{
This function processes mrm_transition for each sample on a plate.
}
\examples{
\dontrun{
process_mrm_transitions(FUNC_mzR, FUNC_mrm_guide, idx_plate, idx_mzML)
}
}
\keyword{internal}
