% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_setup.R
\name{qcCheckR_import_PeakForgeR_reports}
\alias{qcCheckR_import_PeakForgeR_reports}
\title{qcCheckR_import_PeakForgeR_reports}
\usage{
qcCheckR_import_PeakForgeR_reports(master_list)
}
\arguments{
\item{master_list}{The master list object.}
}
\value{
The updated master list object with PeakForgeR reports imported.
}
\description{
This function imports PeakForgeR reports from the project directory and stores them in the master list.
}
\examples{
\dontrun{
master_list <- qcCheckR_import_PeakForgeR_reports(master_list)
}
}
\keyword{internal}
