% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_dataprep.R
\name{qcCheckR_impute_data}
\alias{qcCheckR_impute_data}
\title{Impute Missing Data}
\usage{
qcCheckR_impute_data(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with imputed peak area data.
}
\description{
This function imputes missing and zero values in the \code{master_list} data using the minimum intensity of each feature in the batch divided by 2.
It handles infinite and NaN values, applies imputation, and merges metadata back into the result.
}
\examples{
\dontrun{
master_list <- qcCheckR_impute_data(master_list)
}
}
\keyword{internal}
