% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_filter.R
\name{qcCheckR_lipid_filter}
\alias{qcCheckR_lipid_filter}
\title{Lipid Filter}
\usage{
qcCheckR_lipid_filter(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with lipid filter flags and failed lipid list.
}
\description{
Filters lipids based on missing values across plates and template versions.
Flags lipids with more than 50\% missing values and compiles a list of failed lipids.
}
\examples{
\dontrun{
master_list <- qcCheckR_lipid_filter(master_list)
}
}
\keyword{internal}
