% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_setup.R
\name{qcCheckR_read_mrm_guides}
\alias{qcCheckR_read_mrm_guides}
\title{qcCheckR_read_mrm_guides}
\usage{
qcCheckR_read_mrm_guides(master_list, mrm_template_list)
}
\arguments{
\item{master_list}{The master list object.}

\item{mrm_template_list}{List of MRM guide file paths for tsv.}
}
\value{
The updated master list object with MRM guides.
}
\description{
This function reads MRM guides from user-supplied paths in the mrm_template_list.
}
\examples{
\dontrun{
master_list <- qcCheckR_read_mrm_guides(master_list, mrm_template_list)
}
}
\keyword{internal}
