% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_setup.R
\name{qcCheckR_set_project_details}
\alias{qcCheckR_set_project_details}
\title{qcCheckR_set_project_details}
\usage{
qcCheckR_set_project_details(
  master_list,
  user_name,
  project_directory,
  QC_sample_label,
  sample_tags,
  mv_threshold,
  lod_threshold = 5000
)
}
\arguments{
\item{master_list}{The master list object.}

\item{user_name}{Character string representing the user name for the project.}

\item{project_directory}{Directory path for the project folder.}

\item{QC_sample_label}{Key for filtering QC samples from sample list.}

\item{sample_tags}{Character vector of sample tags to filter sample types from file_names.}

\item{mv_threshold}{Numeric value for the missing value sample threshold.}

\item{lod_threshold}{Numeric instrumental limit of detection (peak area)
below which values are counted as missing. Default is 5000.}
}
\value{
The updated master list object with project details.
}
\description{
This function sets project details in the master list.
}
\examples{
\dontrun{
master_list <- qcCheckR_set_project_details(master_list,
                                            user_name,
                                            project_directory,
                                            QC_sample_label,
                                            sample_tags,
                                            mv_threshold,
                                            lod_threshold)
}
}
\keyword{internal}
