% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_statTarget_batch_correction}
\alias{qcCheckR_statTarget_batch_correction}
\title{Batch Correction and Signal Drift Adjustment}
\usage{
qcCheckR_statTarget_batch_correction(master_list)
}
\arguments{
\item{master_list}{A list containing project details, concentration data, and metadata.}
}
\value{
The updated \code{master_list} with corrected concentration and peak area data.
}
\description{
This function performs batch correction and signal drift adjustment on the concentration data in \code{master_list} using the \code{statTarget} package.
It prepares phenotype and profile files, runs \code{statTarget::shiftCor}, and integrates corrected data back into the master list.
}
\keyword{internal}
