% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{save_plate_data}
\alias{save_plate_data}
\title{save_plate_data}
\usage{
save_plate_data(master_list, plate_idx)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{plate_idx}{The index of the plate to save the data for.}
}
\value{
Saves the master list data to a \code{.qs2} file (load with
\code{\link[qs2:qs_read]{qs2::qs_read()}}) in the specified project directory.
}
\description{
This function saves the master list data for a given plate to a qs2 file.
}
\examples{
\dontrun{
save_plate_data(master_list, plate_idx)
}
}
\keyword{internal}
