% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_docker.R
\docType{data}
\name{SKYLINE_INSTRUMENT_MZ_TOLERANCE}
\alias{SKYLINE_INSTRUMENT_MZ_TOLERANCE}
\title{PeakForgeR_docker.R}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
SKYLINE_INSTRUMENT_MZ_TOLERANCE
}
\description{
Docker command construction, execution, and container management
for running Skyline via Docker.
}
\keyword{datasets}
