% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_archive.R
\name{validate_directories}
\alias{validate_directories}
\title{validate_directories}
\usage{
validate_directories(source_dir, dest_dir)
}
\arguments{
\item{source_dir}{Directory path for the folder to copy.}

\item{dest_dir}{Directory path for the folder to be copied to.}
}
\value{
None. The function performs directory validation.
}
\description{
This function validates the existence of the source directory and creates the destination directory if it does not exist.
}
\examples{
\dontrun{
validate_directories(source_dir = "path/to/source", dest_dir = "path/to/destination")
}
}
\keyword{internal}
