% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_mzml.R
\name{validate_mzR_parameters}
\alias{validate_mzR_parameters}
\title{validate_mzR_parameters}
\usage{
validate_mzR_parameters(FUNC_mzR, FUNC_mrm_guide, FUNC_OPTION_qc_type)
}
\arguments{
\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}

\item{FUNC_mrm_guide}{Tibble of mrm details parsed internally. See run mrm_template_guide for example.}

\item{FUNC_OPTION_qc_type}{QC type used in the experiment (LTR, PQC, none) parsed internally.}
}
\description{
This function validates parameters mzR_mrmfindR. If parameters fail validation script stops
}
\examples{
\dontrun{
validate_mzR_parameters(FUNC_mzR, FUNC_mrm_guide, FUNC_OPTION_qc_type)
}
}
\keyword{internal}
