% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_previous_section}
\alias{validate_previous_section}
\title{Validate Previous Section}
\usage{
validate_previous_section(master_list, previous_section_name)
}
\arguments{
\item{master_list}{A list containing project details and script log information.}

\item{previous_section_name}{A string representing the name of the previous section.}
}
\value{
Called for its side effect; returns \code{NULL} invisibly.
}
\description{
This function validates the previous section name.
}
\keyword{internal}
