% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_MStargetR}
\alias{install_MStargetR}
\title{install_MStargetR}
\usage{
install_MStargetR()
}
\description{
Helper function to install MStargetR. Ensuring all packages
are installed and up to date.
}
\examples{
\dontrun{
install_MStargetR()
}
}
