% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_mzml_conversion}
\alias{msConvertR_mzml_conversion}
\title{msConvertR_mzml_conversion}
\usage{
msConvertR_mzml_conversion(
  input_directory,
  output_directory,
  groups,
  vendor_extension_patterns = MSTARGETR_VENDOR_EXT_PATTERN,
  sanitized_plateIDs = NULL,
  enable_HPC = getOption("MStargetR.enable_HPC", FALSE)
)
}
\arguments{
\item{input_directory}{Directory path for project folder}

\item{output_directory}{Directory path for project folder if different from
input directory.}

\item{groups}{Plate membership table from \code{derive_plate_groups()}.}
}
\value{
Converted mzml files.
}
\description{
This function converts raw vendor files to mzML format using ProteoWizard's msconvert tool, restructures directories, and updates the script log.
}
\examples{
\dontrun{
msConvertR_mzml_conversion(input_directory, output_directory, groups)
}
}
\keyword{internal}
