% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_setup.R
\name{qcCheckR_setup_project}
\alias{qcCheckR_setup_project}
\title{qcCheckR_setup_project}
\usage{
qcCheckR_setup_project(
  user_name,
  project_directory,
  mrm_template_list,
  QC_sample_label,
  sample_tags,
  mv_threshold
)
}
\arguments{
\item{user_name}{Character string representing the user name for the project.}

\item{project_directory}{Directory path for the project folder containing the wiff folder and .wiff and .wiff.scan files for each plate.}

\item{mrm_template_list}{List of lists for mrm_guides.}

\item{QC_sample_label}{Key for filtering QC samples from sample list.}

\item{sample_tags}{Vector of character strings to pull sample types for names}

\item{mv_threshold}{threshold for missing value filter. default is 50\%.}
}
\value{
The updated \code{master_list} object with the project setup details.
}
\description{
This function sets up the project by initialising the master list, setting up project directories, and updating the script log.
}
\examples{
\dontrun{
qcCheckR_setup_project(
  user_name, project_directory, mrm_template_list,
  QC_sample_label, sample_tags, mv_threshold
)
}
}
\keyword{internal}
